requirements.txt
streamlit
streamlit-extras
python-dotenv
google-generativeai